Add ability to send images from server
There are 2 ways to send an image to the browser:

Option 1. Send the image as bytes using OutputStream (which sends bytes, as opposed to Strings). Critical to add the content-type (for example image/png) and the correct content-length (image in bytes) in the headers. The image bytes should be send immediately after the CRLF in between the body and headers -- any extra characters (like an extra CRLF) will corrupt the image bytes being sent through and the browser will not be able to render the data correctly.

Option 2. Read the image as bytes and Base64 Encode it into a string and set it as the SRC for an html image element. For example, sending a html string of <img src='data:image/png;base64,[encodedString]' /> as a response, with the correct content-type text/html will also read correctly in the browser. Advantages of this approach is that we can continue using the PrintWriter output stream (which only handles strings) as our primary writer, and this is also a good case for when we have multiple images stored in a database that needs to be sent through the server -- storing images as a encoded string to send back is much less work than manipulating the image everytime its request.

To get image reading and writing to work, there were a few changes / bugs to be addressed.
- Update Client to use OutputStream (binary data) vs PrintWriter (string). It will continue to keep both options available for the mockClient so that the tests can continue asserting against strings.
- Update FileHandler to add a readFile funtion that reads a file and returns its bytes (so theoretically, this works with all file types including images).
- Update Response to only handle getting/setting response data, with the addition of saving binaryData
- Update ResponseCreator to be ResponseSender. It's job is to grab data from Response and send headers and bodies through the OutputStream. There was a bug in the body creation where it was sending an extra CRLF when trying to send a binary body and thus corrupting the binary data. Update sending of binary responses by manually sending headers, crlf, then the binary data (will refactor later).
- Update Router static content routes to save binary data in the callback
- Update SocketWrapper and its implementations to setOutputStream and a method to use output stream for binary data.