Refactor Design Architecture
Code architecture has been changed to adjust to the following new design requirements: Response is now responsible for creating a full response, which means the Application side has access to new public methods which when called, will create a full valid response. This is to fix the previous logic flow in which the application side was able to make partial responses, and then multiple classes (responseSender and Router) were tasked with validating and filling in the blanks of the Response.

- Update Response to not only hold the response data in parts, but also have public methods to handle invalid requests and creation of the different types of full responses. Finally, response can generate the full response-in-bytes.

- Create ResponseByteAssembler which takes in the response data parts, formats it into a valid HTTP response string, then convert it into bytes

- Remove ResponseSender as the previous logic has been redistributed to associated classes

- Update Router to handle checks for whether a route or method is valid. Router functions have also been moved around to follow a public + associated private implementation functions reading flow.