Fix deletion/directory bugs; Refactor repository
[BUG] File deletion was not working consistently.
[BUG] Directory html was not updated and routable with each new creation and deletion of resources

These 2 bugs were slightly correlated.

This had to do with GET/DELETE methods not being implemented for the different route options. For example, a new resource of "/dog/1" should be able to have GET and DELETE methods for "/dog/1", "dog/1.txt", "/public/dog/1", and "/public/dog/1.txt". This allows users to get from a route they expect, and for the directory html page to route correctly to the file as well.

Another issue was trying to save a resource as a MIMEType/Filetype that is not established in my MIMETypes data structure, which then creates a null file type. Any attempts to get or delete this file is then met with server errors as the expected file was not created, and now there is no link to the file.

- Update Repository interface to only include the public methods that both Repository types need to expose.
- Update MockReposiory and its usage. The MockRepository is now instantiated before each test. Instead of having a default set of files on MockRepository instantiation, it is now created as an empty "directory" from which `writeFile` can be used to add files.
- Update MockRouter to require a repository on instantiation.
- Add RouterTest to check for correct resource creation and deletion