Update Server to handle HTML Form data in Request
HTML Form data can be taken care of in 2 ways:
(1) If the html form was defined with a GET method, the route that handles the request needs to parse form data from the Request route (form data is set as parameters in the route).
(2) If the html form was defined with a POST method, the route that handles the request needs to parse form data from the Request body. It should be able to handle a html form defined enctype of "application/x-www-form-urlencoded"

- Update README with Quick Start guide and public methods for Router, Request and Response