Fix bug where Server fails when form is submitted
Issue: When on route /post_form, the first form data gets submitted to server correctly. However, when trying to submit another form, the Request never completes and the server has an error.

After looking at the server logs, the breaking change was from router.findGreatestResourceId(). Its function is to find the next greatest resource ID -- for example, if /dogs/1 exists, it should return /dogs/2 as the next available route the new resource can be saved under. However, it was not ready to handle routes that aren't parsable numbers. In this case, we tried to pass in the route "/post_form/1.txt", and it could not parse the ".txt" portion of the string and thus it failed.

The fix is to add a check for routes with file types (for example, ".txt") in the name so it can grab just the resource name.