Add error handling to Server operations
Issue: Trying to run a patch operation  ("remove", "update") on a non-existent JSON Key should send back an error. This is also a good time to group the other common patch errors and associated status codes within the error handling.

After reviewing the rest of the code, it looks like error handling is also missing in the other server operations like POST.

Fix: To accomplish this, we need to create a error handling strategy to bubble up the errors to the Response class.  The current plan is to use a new OperationResult class to keep track of operation results, which is then passed into the Response methods to generate a "Error" Response or a "Everything Worked" Response

- Create JSONPatcher class to handle validation then running of the Patch document. Both functions will now return OperationResult type
- Create OperationResult class which is used after Server tries to read/write/update a resource. It holds whether the operation was valid, status code, and result of operation data.
- Update ResourceHandler. Rename to RouterResourceHandler, remove private members, and make it clear that there is a tight coupling to Router.
- Update Router to move all resource + routes for resource handling into RouterResourceHandler
- Create RouterDirectoryHandler from RouterResourceHandler. This class handles the creation and usage of the directory route
-  Update Response methods to use OperationResult